test - make time cast to integer

+++ tests/euler/test.java
@@ -224,7 +224,7 @@ public class test {
             assertEquals(result, cls);
             double end = System.currentTimeMillis();
             double time = end - start;
-            if (time > logTime) System.out.println(cls.getName() + ": " + time + " ms");
+            if (time > logTime) System.out.println(cls.getName() + ": " + (int) time + " ms");
         }
         double allEnd = System.currentTimeMillis();
         System.out.println("Total time: " + (allEnd - allStart) + " ms");
